Add files via upload
This is a calculator I developed out of a practice problem from an algorithm book. Started with pseudocode and desk-checking (the math equations worked!) but had some trouble debugging syntax issues for many hours (likely due to some unfamiliar aspects of C). One of those problems that was satisfying once the solution came together. Feel free to use or adopt, it was excellent practice for using If-Else cases and implementing some exceptions!